+++ packages/@leight/file/src/sdk/Source.ts
@@ -19,4 +19,4 @@ export const FileSourceContext                = (container: IContainer) => new S
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_p15q1rqzh8kmonetfg70cnyq = true;
+export const $leight_ju3e69gqog3yllgttr529epl = true;