+++ packages/@leight/job/src/sdk/SourceSchema.ts
@@ -34,4 +34,4 @@ export interface IJobSourceSchema extends ISourceSchema<
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_ptqztmskk842224tqb1tdcfz = true;
+export const $leight_nh5vqm65tsbitl661wxfyqwu = true;