+++ packages/@leight/user/src/sdk/PrismaSchema.ts
@@ -71,4 +71,4 @@ export const UserQuerySchema                  = QuerySchema({
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_a35egtc4vlwgem2fmni6jyvh = true;
+export const $leight_kwbzdqo135exmcjwesvzesq4 = true;