+++ packages/@leight/context/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @leight/context
 
+## 0.5.25
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/zustand@0.5.62
+
 ## 0.5.24
 
 ### Patch Changes