+++ packages/@leight/esbuild/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leight/esbuild
 
+## 0.5.41
+
+### Patch Changes
+
+- Version bump
+
 ## 0.5.40
 
 ### Patch Changes