+++ packages/@leight/filter/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @leight/filter
 
+## 0.5.58
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/context@0.5.25
+    - @leight/zod@0.5.8
+    - @leight/zustand@0.5.62
+
 ## 0.5.57
 
 ### Patch Changes