+++ packages/@leight/generator-server/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @leight/generator-server
 
+## 0.5.18
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/generator@0.5.17
+    - @leight/utils@0.5.65
+    - @leight/utils-server@0.5.61
+
 ## 0.5.17
 
 ### Patch Changes