+++ packages/@leight/mantine/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @leight/mantine
 
+## 0.5.69
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/file-client@0.5.71
+    - @leight/file@0.5.70
+    - @leight/i18n@0.5.68
+    - @leight/i18n-client@0.5.68
+    - @leight/layout@0.5.62
+    - @leight/utils@0.5.65
+    - @leight/utils-client@0.5.66
+
 ## 0.5.68
 
 ### Patch Changes