+++ packages/@leight/query/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @leight/query
 
+## 0.5.62
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/cursor@0.5.56
+    - @leight/filter@0.5.58
+    - @leight/react-query@0.5.25
+    - @leight/sort@0.5.59
+    - @leight/zod@0.5.8
+
 ## 0.5.61
 
 ### Patch Changes