+++ packages/@leight/sdk/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @leight/sdk
 
+## 0.5.20
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/generator@0.5.17
+    - @leight/generator-server@0.5.18
+    - @leight/utils@0.5.65
+    - @leight/utils-server@0.5.61
+
 ## 0.5.19
 
 ### Patch Changes