+++ packages/@leight/source-client/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @leight/source-client
 
+## 0.5.33
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/cursor-client@0.5.32
+    - @leight/context@0.5.25
+    - @leight/context-client@0.5.70
+    - @leight/filter@0.5.58
+    - @leight/sort@0.5.59
+    - @leight/source@0.5.62
+    - @leight/utils@0.5.65
+    - @leight/zustand@0.5.62
+
 ## 0.5.32
 
 ### Patch Changes