+++ packages/@leight/sql/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @leight/sql
 
+## 0.5.62
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/prisma@0.5.79
+
 ## 0.5.61
 
 ### Patch Changes