+++ packages/@leight/zustand/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @leight/zustand
 
+## 0.5.62
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/utils@0.5.65
+
 ## 0.5.61
 
 ### Patch Changes